Fix SCAN command cursor overflow with __uint128_t and optimize code reusability
- Use __uint128_t to handle large cursor values from Kvrocks (up to 64-bit)
- Prevent integer overflow when encoding cursor with Group ID
- Add Util::uint128ToString() utility function to eliminate code duplication
- Refactor Handler.cpp and Request.cpp to use the new utility function
- Add debug logging for cursor transformation tracking

This fix resolves SCAN cursor infinite loop issue when using Predixy with Kvrocks in Cluster mode.

+++ src/Request.h
@@ -63,7 +63,7 @@ public:
     void setSentinels(const String& master);
     void setSentinelGetMaster(const String& master);
     void setSentinelSlaves(const String& master);
-    void adjustScanCursor(long cursor);
+    void adjustScanCursor(__uint128_t cursor);
     void follow(Request* leader);
     void setResponse(Response* res);
     bool send(Socket* s);
